proper operator frequencies

+++ synth/src/fm.h
@@ -31,7 +31,7 @@ typedef struct {
 
 typedef struct {
     uint8_t algorithm;
-    double freq_ratios[FM_OP_COUNT];
+    uint8_t freq_ratios[FM_OP_COUNT];
     double amplitudes[FM_OP_COUNT];
 
     uint8_t feedback_type;
